children with 2 context demo
