+++ components/modal/demo/hooks.md
@@ -8,11 +8,11 @@ only: true
 
 ## zh-CN
 
-通过 `Modal.useModal` 调取支持 ConfigProvider context 的对话框。组件上层必须要有 ConfigProvider 组件。
+通过 `Modal.useModal` 创建支持读取 context 的 `contextHolder`。
 
 ## en-US
 
-Use `Modal.useModal` to call with ConfigProvider context support modal. ConfigProvider must exist in parent.
+Use `Modal.useModal` to get `contextHolder` with context accessible issue.
 
 ```jsx
 import { Modal, Button } from 'antd';
@@ -28,38 +28,42 @@ const config = {
   },
 };
 
-const Context = React.createContext();
+const ReachableContext = React.createContext();
+const UnreachableContext = React.createContext();
 
 const App = () => {
-  const [modal] = Modal.useModal();
-  const [inlineModal, inlineHolder] = Modal.useModal();
+  const [modal, contextHolder] = Modal.useModal();
 
   return (
-    <div>
+    <ReachableContext.Provider value="Light">
       <Button
         type="primary"
         onClick={() => {
-          modal.confirm(config);
+          modal.confirm({
+            ...config,
+            content: (
+              <div>
+                <ReachableContext.Consumer>
+                  {name => `Reachable: ${name}!`}
+                </ReachableContext.Consumer>
+                <br />
+                <UnreachableContext.Consumer>
+                  {name => `Unreachable: ${name}!`}
+                </UnreachableContext.Consumer>
+              </div>
+            ),
+          });
         }}
       >
-        Use ConfigProvider
+        Use inline holder
       </Button>
 
-      <Context.Provider value="Light">
-        <Button
-          type="primary"
-          onClick={() => {
-            inlineModal.confirm({
-              ...config,
-              content: <Context.Consumer>{name => `Hello, ${name}!`}</Context.Consumer>,
-            });
-          }}
-        >
-          Use inline holder
-        </Button>
-        {inlineHolder}
-      </Context.Provider>
-    </div>
+      {/* `contextHolder` should always under the context you want to access */}
+      {contextHolder}
+
+      {/* Can not access this context since `contextHolder` is not in it */}
+      <UnreachableContext.Provider value="Bamboo" />
+    </ReachableContext.Provider>
   );
 };
